Доделал проверки по комментариям
