+++ js/utils.js
@@ -1,4 +1,6 @@
 const bodyElement = document.querySelector('body');
+const formElement = document.querySelector('.img-upload__form');
+const hastagTextElement = formElement.querySelector('.text__hashtags');
 
 const getId = () => {
   let lastGeneratedId = 0;
@@ -27,4 +29,4 @@ const disableEscEvt = (element) => {
   });
 };
 
-export {getId, getRandomInteger, getRandomArrayElement, isEscapeKey, toggleModalOpen, disableEscEvt};
+export {getId, getRandomInteger, getRandomArrayElement, isEscapeKey, toggleModalOpen, disableEscEvt, formElement, hastagTextElement};
